+++ operator/internal/resources/aisauth/doc.go
@@ -1,0 +1,6 @@
+/*
+ * Copyright (c) 2026, NVIDIA CORPORATION. All rights reserved.
+ */
+
+// Package aisauth contains Kubernetes resources for the AuthN authentication server.
+package aisauth